requirements.txt
numpy
pandas
scikit-learn
nltk
spacy
PyPDF2
python-docx
matplotlib
seaborn
streamlit
sentence-transformers